Update DDA JSON, obsolete ported over content
* JSON updates to waist item and itemgroups.
* Obsoleted close helm because @QuartzMech added more or less the same idea in https://github.com/CleverRaven/Cataclysm-DDA/pull/49985

I was tempted to retain the kettle hat item and recipe as a full override because the DDA version is pretty badly implemented, lacking the mod version's idea of actually filling the niche of a decent autolearned iron-tier helmet, but eh.

Also seriously, make it overtly a wide-brimmed iron helmet, emphasize it has a wide brim, but deliberately NOT make it shade the eyes? My version had that to begin with.